requirements_paddle.txt
paddlepaddle==2.5.2
paddleocr==2.7.3
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
numpy==1.24.3 